fix(ddei): support Linux 386 and release all three architectures

+++ internal/detector/ddeirootkit/ctime_linux.go
@@ -13,5 +13,5 @@ func ctimeOf(path string) time.Time {
 		return time.Time{}
 	}
 	sec, nsec := st.Ctim.Sec, st.Ctim.Nsec
-	return time.Unix(sec, nsec)
+	return time.Unix(int64(sec), int64(nsec))
 }
